requirements.txt
google-api-python-client == 2.102.0
google-auth == 2.27.0
google-cloud-bigquery == 3.11.4
pandas == 1.4.4
numpy == 1.24.4
openai == 1.41.1
langchain-text-splitters == 0.2.2
tqdm == 4.64.1
python-dotenv == 1.0.0
Janome ==  0.5.0
rank-bm25 == 0.2.2
vertexai == 1.68.0